+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Internal**:
 
 - Move unreal crash report expansion from processing into endpoint. ([#5825](https://github.com/getsentry/relay/pull/5825))
+- Retry failing objectstore requests. ([#5836](https://github.com/getsentry/relay/pull/5836))
 
 ## 26.4.0